Melhoramento da organização do código
